:truck: move websocket module
