+++ webhooks/dependencies/sse_manager.py
@@ -16,7 +16,7 @@ from shared.log_config import get_logger
 from shared.models.webhook_topics import WEBHOOK_TOPIC_ALL, CloudApiWebhookEventGeneric
 from webhooks.dependencies.event_generator_wrapper import EventGeneratorWrapper
 from webhooks.dependencies.redis_service import RedisService
-from webhooks.dependencies.websocket import publish_event_on_websocket
+from webhooks.routers.websocket import publish_event_on_websocket
 
 logger = get_logger(__name__)
